Added levelSet table
