+++ backend/src/main/python/utils/insert_levels.py
@@ -2,7 +2,6 @@ from collections import defaultdict
 import requests
 import random
 
-
 def insert_levels(hasura_url, headers, editions, random_gen, max_points_in_level, levels_data):
     def generate_levels():
         level_steps = [0]
@@ -42,18 +41,23 @@ def insert_levels(hasura_url, headers, editions, random_gen, max_points_in_level
             }
             """
             variables_file = {"filename": filename}
-            response_file = requests.post(
-                hasura_url,
-                json={"query": query_file_id, "variables": variables_file},
-                headers=headers
-            )
-            file_data = response_file.json()
-
-            if "errors" in file_data or not file_data.get("data", {}).get("files"):
-                print(f"Error fetching file ID for '{filename}': {file_data.get('errors', 'File not found')}")
-                image_file_id = None  # or handle as per your requirements
+            try:
+                response_file = requests.post(
+                    hasura_url,
+                    json={"query": query_file_id, "variables": variables_file},
+                    headers=headers
+                )
+                response_file.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                print(f"HTTP error while fetching file ID for '{filename}': {e}")
+                image_file_id = None
             else:
-                image_file_id = file_data["data"]["files"][0]["fileId"]
+                file_data = response_file.json()
+                if "errors" in file_data or not file_data.get("data", {}).get("files"):
+                    print(f"Error fetching file ID for '{filename}': {file_data.get('errors', 'File not found')}")
+                    image_file_id = None  # or handle as per your requirements
+                else:
+                    image_file_id = file_data["data"]["files"][0]["fileId"]
 
             # Prepare the LevelInputType dictionary
             level_input = {
@@ -64,14 +68,18 @@ def insert_levels(hasura_url, headers, editions, random_gen, max_points_in_level
             }
             level_inputs.append(level_input)
 
+        level_set_name = ""
+
         # Mutation to add a level set
         mutation_add_level_set = """
         mutation AddLevelSet($levels: [LevelInputType!]!) {
             addLevelSet(levels: $levels) {
-                levelSet
-                levelId
-                levelName
-                ordinalNumber
+                levelSetId
+                levelSetName
+                levels {
+                    levelId
+                    ordinalNumber
+                }
             }
         }
         """
@@ -79,28 +87,30 @@ def insert_levels(hasura_url, headers, editions, random_gen, max_points_in_level
             "levels": level_inputs
         }
 
-        response_add_level_set = requests.post(
-            hasura_url,
-            json={"query": mutation_add_level_set, "variables": variables_add_level_set},
-            headers=headers
-        )
+        try:
+            response_add_level_set = requests.post(
+                hasura_url,
+                json={"query": mutation_add_level_set, "variables": variables_add_level_set},
+                headers=headers
+            )
+            response_add_level_set.raise_for_status()
+        except requests.exceptions.RequestException as e:
+            print(f"HTTP error while adding level set for edition {edition_id}: {e}")
+            continue  # Skip to the next edition
+
         data_add_level_set = response_add_level_set.json()
 
         if "errors" in data_add_level_set:
             print(f"Error adding level set for edition {edition_id}: {data_add_level_set['errors']}")
             continue  # Skip to the next edition
 
-        added_levels = data_add_level_set["data"]["addLevelSet"]
-        print(f"Successfully added {len(added_levels)} levels for edition {edition_id}")
+        added_level_set = data_add_level_set["data"]["addLevelSet"]
+        print(f"Successfully added level set '{added_level_set['levelSetName']}' with ID {added_level_set['levelSetId']} for edition {edition_id}")
 
         # Collect level IDs and ordinals
-        inserted_levels[edition_id] = [(level["levelId"], level["ordinalNumber"]) for level in added_levels]
+        inserted_levels[edition_id] = [(level["levelId"], level["ordinalNumber"]) for level in added_level_set["levels"]]
 
-        # Assuming that addLevelSet returns a unique levelSet ID, which might need to be fetched differently
-        # If not, you might need to modify the backend to return it or generate it here
-        # For demonstration, let's assume levelSet ID is the first level's ID (this may not be accurate)
-        # You should replace this with the correct logic to obtain the levelSet ID
-        level_set_id = added_levels[0]["levelSet"] if added_levels else None
+        level_set_id = added_level_set["levelSetId"]
 
         if not level_set_id:
             print(f"Unable to determine levelSet ID for edition {edition_id}. Skipping association.")
@@ -108,29 +118,40 @@ def insert_levels(hasura_url, headers, editions, random_gen, max_points_in_level
 
         # Mutation to assign the level set to the edition
         mutation_assign_level_set = """
-        mutation AddLevelSetToEdition($levelSet: Int!, $editionId: Int!) {
-            addLevelSetToEdition(levelSet: $levelSet, editionId: $editionId) {
-                levelId
-                levelName
-                ordinalNumber
+        mutation AddLevelSetToEdition($levelSetID: Int!, $editionId: Int!) {
+            addLevelSetToEdition(levelSetID: $levelSetID, editionId: $editionId) {
+                levelSetId
+                levelSetName
+                edition {
+                    editionId
+                }
+                levels {
+                    levelId
+                    ordinalNumber
+                }
             }
         }
         """
         variables_assign_level_set = {
-            "levelSet": level_set_id,
+            "levelSetID": level_set_id,
             "editionId": edition_id
         }
 
-        response_assign_level_set = requests.post(
-            hasura_url,
-            json={"query": mutation_assign_level_set, "variables": variables_assign_level_set},
-            headers=headers
-        )
+        try:
+            response_assign_level_set = requests.post(
+                hasura_url,
+                json={"query": mutation_assign_level_set, "variables": variables_assign_level_set},
+                headers=headers
+            )
+            response_assign_level_set.raise_for_status()
+        except requests.exceptions.RequestException as e:
+            print(f"HTTP error while assigning level set {level_set_id} to edition {edition_id}: {e}")
+            continue
+
         data_assign_level_set = response_assign_level_set.json()
 
         if "errors" in data_assign_level_set:
-            print(
-                f"Error assigning level set {level_set_id} to edition {edition_id}: {data_assign_level_set['errors']}")
+            print(f"Error assigning level set {level_set_id} to edition {edition_id}: {data_assign_level_set['errors']}")
         else:
             print(f"Successfully assigned level set {level_set_id} to edition {edition_id}")
